+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - ConnectionFactory.createConnection() and ConnectionProvider.connection() overload without User parameter added.
 ### is.codion.common.model
 - FilterModel.Refresher.success() and failure() renamed result() and exception() respectively, related renaming.
+- FilterModel.Items no longer extends Observable.
 ### is.codion.framework.domain
 - DefaultEntityValidator.valid() now final.
 - AttributeDefinition.Builder.mnemonicResourceKey() added.